LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/mo-stretch-properties-dynamic-001-expected.txt


FAIL minsize assert_approx_equals: attach expected 100 +/- 1 but got 0
FAIL maxsize assert_approx_equals: attach expected 100 +/- 1 but got 0
FAIL largeop assert_approx_equals: set true expected 50 +/- 1 but got 17
FAIL symmetric assert_approx_equals: set true expected 150 +/- 1 but got 0
FAIL stretchy assert_approx_equals: set true expected 75 +/- 1 but got 0
⥯
⥯
⥯

⥯
⥯
⥯

⥯
⥯

⥯
⥯

⥯
⥯
